// demo-ui/src/app/features/generated/dummy/stupid/models/StupidBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class StupidBasicView {

id: number;
selected: boolean;
canDelete: boolean;
code: string;
foolCode: string;

/* Specific Code Start */
/* Specific Code End */
}