// demo-ui/src/app/features/generated/dummy/stupid/list/modal/stupid-modal.component.ts
import { Component, OnInit, ViewChild } from '@angular/core';
import { FoolRestClient } from 'src/app/features/generated/dummy/fool/services/FoolRestClient';
import { FormBuilder, FormGroup, Validators } from '@angular/forms';
import { MatDialogRef } from '@angular/material/dialog';
import { NotificationService } from 'src/app/core/services/NotificationService';
import { Observable } from 'rxjs';
import { SelectItem } from 'src/app/core/models/SelectItem';
import { StringUtils } from 'src/app/core/services/StringUtils';
import { StupidFullView } from '../../models/StupidFullView';
import { StupidRestClient } from '../../services/StupidRestClient';
/**
 * auto generated modal component ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */

@Component({
selector: 'app-stupid-modal',
templateUrl: './stupid-modal.component.html',
styleUrls: ['./stupid-modal.component.scss']
})
export class StupidModalComponent implements OnInit {

view: StupidFullView;
form: FormGroup;
foolCodeOptions: Observable<SelectItem[]>;

constructor(private service:StupidRestClient, private foolService:FoolRestClient, private formBuilder: FormBuilder, private notifications: NotificationService, private dialogRef: MatDialogRef<StupidModalComponent>) {}

ngOnInit(): void {
this.form = this.formBuilder.group({
code:[null, Validators.required],
foolCode:[null, Validators.required]});
this.form.controls['foolCode'].valueChanges.subscribe(value=>{this.searchOptionsForFoolCode(value)});
this.restoreForm();
}
restoreForm(): void {
this.form.patchValue({
code: this.view.form.code,
foolCode: this.view.form.foolCode
})
}

applyForm(): void {
this.view.form.code = StringUtils.emptyToNull(this.form.get('code').value);
this.view.form.foolCode = StringUtils.emptyToNull(this.form.get('foolCode').value);
}

searchOptionsForFoolCode(value:string) {
if (value) {
this.foolCodeOptions = this.foolService.searchOptions(value);
}
}

save(): void {
this.applyForm();
this.service.save(this.view.form).subscribe(success => {this.notifications.info("Operation completed");this.dialogRef.close();}, error => {this.notifications.error("Operation failed")});
}

update(): void {
this.applyForm();
this.service.update(this.view.id, this.view.form).subscribe(success => {this.notifications.info("Operation completed");this.dialogRef.close();}, error => {this.notifications.error("Operation failed")});
}

saveOrUpdate(): void {
if (this.view.id == null) {this.save()} else {this.update()}
}

/* Specific Code Start */
/* Specific Code End */
}